+++ CHANGELOG.md
@@ -12,6 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Complete Portuguese translations
 - A mechanism to translate quote status into multiple languages
 
+## [1.2.0] - 2022-06-28
+
+### Added
+
+- UX adjustments on Alert messages and action buttons positions.
+- It was reorganized the code and split components into small pieces.
+
+
 ## [1.1.1] - 2022-06-06
 
 ### Fixed